Revert "introduce hook"
This reverts commit 9bb6c2836cf1e62310774c879b89ad769f68dfe6.
