+++ static/app/views/projectInstall/platformDocHeader.tsx
@@ -1,4 +1,4 @@
-import {useCallback} from 'react';
+import {useCallback, useMemo} from 'react';
 import {useBlocker} from 'react-router-dom';
 import styled from '@emotion/styled';
 
@@ -6,7 +6,7 @@ import {removeProject} from 'sentry/actionCreators/projects';
 import {Button} from 'sentry/components/core/button';
 import {ButtonBar} from 'sentry/components/core/button/buttonBar';
 import {LinkButton} from 'sentry/components/core/button/linkButton';
-import {useProjectState} from 'sentry/components/onboarding/useProjectState';
+import {useRecentCreatedProject} from 'sentry/components/onboarding/useRecentCreatedProject';
 import type {Platform} from 'sentry/data/platformPickerCategories';
 import {IconChevron} from 'sentry/icons/iconChevron';
 import {t} from 'sentry/locale';
@@ -14,6 +14,7 @@ import {space} from 'sentry/styles/space';
 import type {Project} from 'sentry/types/project';
 import {trackAnalytics} from 'sentry/utils/analytics';
 import {handleXhrErrorResponse} from 'sentry/utils/handleXhrErrorResponse';
+import {isProjectActive} from 'sentry/utils/projects';
 import type RequestError from 'sentry/utils/requestError/requestError';
 import useApi from 'sentry/utils/useApi';
 import useOrganization from 'sentry/utils/useOrganization';
@@ -31,7 +32,24 @@ export function PlatformDocHeader({platform, title, project}: Props) {
   const api = useApi({persistInFlight: true});
   const router = useRouter();
 
-  const projectState = useProjectState(project);
+  const recentCreatedProject = useRecentCreatedProject({
+    orgSlug: organization.slug,
+    projectSlug: project.slug,
+  });
+
+  const projectState = useMemo(() => {
+    if (recentCreatedProject.project) {
+      return {
+        project: recentCreatedProject.project,
+        active: recentCreatedProject.isProjectActive,
+      };
+    }
+
+    return {
+      project,
+      active: isProjectActive(project),
+    };
+  }, [recentCreatedProject.project, recentCreatedProject.isProjectActive, project]);
 
   const handleGoBack = useCallback(async () => {
     trackAnalytics('project_creation.back_button_clicked', {
@@ -41,23 +59,23 @@ export function PlatformDocHeader({platform, title, project}: Props) {
     if (!projectState.active) {
       trackAnalytics('project_creation.data_removal_modal_confirm_button_clicked', {
         organization,
-        platform: projectState.slug,
-        project_id: projectState.id,
+        platform: projectState.project.slug,
+        project_id: projectState.project.id,
       });
 
       try {
         await removeProject({
           api,
           orgSlug: organization.slug,
-          projectSlug: projectState.slug,
+          projectSlug: projectState.project.slug,
           origin: 'getting_started',
         });
 
         trackAnalytics('project_creation.data_removed', {
           organization,
-          date_created: projectState.dateCreated,
-          platform: projectState.slug,
-          project_id: projectState.id,
+          date_created: projectState.project.dateCreated,
+          platform: projectState.project.slug,
+          project_id: projectState.project.id,
         });
       } catch (error) {
         handleXhrErrorResponse(
@@ -72,7 +90,7 @@ export function PlatformDocHeader({platform, title, project}: Props) {
       makeProjectsPathname({
         path: '/new/',
         organization,
-      }) + `?referrer=getting-started&project=${projectState.id}`
+      }) + `?referrer=getting-started&project=${projectState.project.id}`
     );
   }, [api, projectState, organization, router]);
